Move set_part_sizes() declaration to status_bar.h
