+++ foo_ui_columns/config_items.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 
 #include "main_window.h"
+#include "status_bar.h"
 
 namespace main_window {
 const char* default_status_pane_script
