+++ foo_ui_columns/main_window.h
@@ -41,19 +41,6 @@ enum {
     MSG_NOTICATION_ICON
 };
 
-namespace status_bar {
-enum t_parts : uint32_t {
-    t_parts_none = 0,
-    t_parts_all = 0xffffffff,
-    t_part_main = 1 << 0,
-    t_part_lock = 1 << 1,
-    t_part_length = 1 << 2,
-    t_part_volume = 1 << 3
-};
-
-void set_part_sizes(unsigned p_parts = t_parts_none);
-}; // namespace status_bar
-
 bool remember_window_pos();
 
 void update_status();
