+++ foo_ui_columns/status_bar.h
@@ -23,6 +23,17 @@ extern HTHEME thm_status;
 
 extern volume_popup_t volume_popup_window;
 
+enum t_parts : uint32_t {
+    t_parts_none = 0,
+    t_parts_all = 0xffffffff,
+    t_part_main = 1 << 0,
+    t_part_lock = 1 << 1,
+    t_part_length = 1 << 2,
+    t_part_volume = 1 << 3
+};
+
+void set_part_sizes(unsigned p_parts = t_parts_none);
+
 HICON get_icon();
 void destroy_icon();
 void destroy_theme_handle();
